.claude/skills/rss-article-saver/pyproject.toml
[project]
name = "rss-article-saver"
version = "1.0.0"
description = "RSS article subscription with AI scoring, save high-quality articles to local Markdown and Notion"
readme = "SKILL.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.0",
    "PyYAML>=6.0",
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "openai>=1.0.0",
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
